% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_kinship_data.R
\name{family_dataset_generator}
\alias{family_dataset_generator}
\title{family_dataset_generator}
\usage{
family_dataset_generator(
  n_clusters = 100,
  n_individuals_in_cluster = 10,
  variance_param = c("R", "H", "K"),
  n_mean_param = 1,
  beta_0 = 1,
  beta_1 = 3,
  beta_2 = 3,
  sigma_R = 1,
  sigma_H = 3,
  sigma_K = 5,
  seed = 1
)
}
\arguments{
\item{n_clusters}{Integer. Number of clusters.}

\item{n_individuals_in_cluster}{Integer. Number of individuals in given cluster.}

\item{variance_param}{Vector. Vector of characters taking values 'R', 'H' and/or 'K' specifying covariance structure.}

\item{n_mean_param}{Integer. Specify number of fixed effects. Taking values 1,2,3.}

\item{beta_0}{Real. Fixed effect 1.}

\item{beta_1}{Real. Fixed effect 2.}

\item{beta_2}{Real. Fixed effect 3.}

\item{sigma_R}{Real. Residual variance.}

\item{sigma_H}{Real. Household variance.}

\item{sigma_K}{Real. Kinship variance.}

\item{seed}{NA or integer. If left unspecified no seed is set. Otherwise set a seed.}
}
\value{
List. Returns a list consisting of a list of outcomes, list of gamma-matrices and list of design matrices
}
\description{
function that generates data based on the genetic matrices residual, household, kinship
}
