% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_kinship_data.R
\name{FDG}
\alias{FDG}
\title{Family data set generator object}
\usage{
FDG(
  n_clusters = 100,
  n_individuals = 10,
  variance_param = c("R", "H", "K"),
  n_mean_param = 1,
  beta_0 = 1,
  beta_1 = 3,
  beta_2 = 3,
  sigma_R = 1,
  sigma_H = 3,
  sigma_K = 5,
  seed = NA
)
}
\arguments{
\item{n_clusters}{an integer denoting the number of clusters.}

\item{n_individuals}{an integer denoting the number of individuals in given cluster.}

\item{variance_param}{a vector of characters taking values 'R', 'H' and/or 'K' specifying covariance structure.}

\item{n_mean_param}{an integer specifying number of fixed effects. Taking values 1,2,3.}

\item{beta_0}{a real denoting the first fixed effect.}

\item{beta_1}{a real denoting the second fixed effect.}

\item{beta_2}{a real denoting the third fixed effect.}

\item{sigma_R}{a real denoting the residual variance.}

\item{sigma_H}{a real denoting the household variance.}

\item{sigma_K}{a real denoting the kinship variance.}

\item{seed}{NA or integer. If left unspecified no seed is set. Otherwise set a seed.}
}
\value{
Family data set
}
\description{
Object for generating genetics data.
}
