% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_kinship_data.R
\name{generate_kinship_matrix}
\alias{generate_kinship_matrix}
\title{generate_kinship_matrix}
\usage{
generate_kinship_matrix(n_individuals, seed = NA)
}
\arguments{
\item{n_individuals}{Integer. Number of individuals in given cluster}

\item{seed}{NA or integer. If left unspecified no seed is set. Otherwise set a seed.}
}
\value{
Matrix. Returns a random kinship matrix of dimensions n_individuals x n_individuals
}
\description{
function that generate kinship matrix
}
