% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_helper_functions.R
\name{get_outcome_variance}
\alias{get_outcome_variance}
\title{Title}
\usage{
get_outcome_variance(outcome)
}
\arguments{
\item{outcome}{list of list of outcomes.}
}
\value{
the variance of the list of lists of outcomes.
}
\description{
Title
}
