% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_params.R
\name{init_params}
\alias{init_params}
\title{Initialize parameters}
\usage{
init_params(design_matrices, outcome_list)
}
\arguments{
\item{design_matrices}{list of design matrices}

\item{outcome_list}{list of outcomes}
}
\value{
initial parameters for the fisher scoring algorithm
}
\description{
Initialize parameters
}
