% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fisher_scoring.R
\name{lmmgen}
\alias{lmmgen}
\title{Find MLE parameters}
\usage{
lmmgen(
  init_fixed_effect_par = NULL,
  init_random_effect_par = NULL,
  design_matrices,
  semi_def_matrices,
  outcome_list,
  update_step_size = 1,
  max_iter = 100,
  tolerance = 1e-12,
  small_value_threshold = 1e-09,
  add_small_constant = 1e-12
)
}
\arguments{
\item{init_fixed_effect_par}{vector of initial fixed effects parameters. Make sure the order of the initial fixed effects match the order of the columns in the design matrices.}

\item{init_random_effect_par}{vector of initial random effects parameters. Make sure the order of the initial random effects match the order of the variance/covariance effects in the semi_def_matrices.}

\item{design_matrices}{list of design matrices for fixed effects for each block.}

\item{semi_def_matrices}{list of lists of variance/covariance matrices.}

\item{outcome_list}{list of vectors of outcomes for each block.}

\item{update_step_size}{update steps are multiplied with this value.}

\item{max_iter}{an integer determining the maximum number of iterations.}

\item{tolerance}{numeric value to determine convergence tolerance.}

\item{small_value_threshold}{values in matrices below this threshold is set to 0.}

\item{add_small_constant}{adds this constant to diagonal of matrices with small values for stability.}
}
\value{
The parameter estimates, the inverse fisher information, the score and a summary.
}
\description{
Finds the MLE parameters using the fisher scoring algorithm.
}
