% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Log-likelihood}
\usage{
log_likelihood(
  design_matrices,
  semi_def_matrices,
  outcome_list,
  fixed_effect_par,
  random_effect_par
)
}
\arguments{
\item{design_matrices}{list of fixed effects design matrices - one for each block.}

\item{semi_def_matrices}{list of lists of covariance structure matrices for each block.}

\item{outcome_list}{list of vector of outcomes for each block.}

\item{fixed_effect_par}{vector of fixed effect parameters.}

\item{random_effect_par}{vector of random effect parameters.}
}
\value{
The log-likelihood.
}
\description{
Log-likelihood
}
