% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood_block}
\alias{log_likelihood_block}
\title{Block-wise log-likelihood}
\usage{
log_likelihood_block(
  design_matrix,
  semi_def_matrix,
  outcomes,
  fixed_effect_par,
  random_effect_par
)
}
\arguments{
\item{design_matrix}{fixed effects design matrix for a single block.}

\item{semi_def_matrix}{list of covariance structure matrices for each block.}

\item{outcomes}{vector of outcomes for the block.}

\item{fixed_effect_par}{vector of fixed effect parameters.}

\item{random_effect_par}{vector of random effect parameters.}
}
\value{
The log-likelihood for a single block.
}
\description{
Block-wise log-likelihood
}
