% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_func.R
\name{omega_func}
\alias{omega_func}
\title{omega_func}
\usage{
omega_func(semi_def_matrix, sigma2_vec)
}
\arguments{
\item{semi_def_matrix}{A list containing the three matrices, householder matrix, kinship matrix, identity matrix.}

\item{sigma2_vec}{A vector containing the variances of the random effects in the same order as the three matrices.}
}
\value{
The variance/covariance matrix Omega
}
\description{
Calculates the variance/covariance matrix Omega or V.
}
\examples{
mat1 <- matrix(1:9, ncol = 3)
mat2 <- matrix(10:18, ncol = 3)
mat3 <- matrix(19:27, ncol = 3)
omega_func(semi_def_matrix = list(mat1, mat2, mat3), sigma2_vec = c(1,2,3))
}
