% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_kinship_data.R
\name{pair_children_with_spouses}
\alias{pair_children_with_spouses}
\title{pair_children_with_spouses}
\usage{
pair_children_with_spouses(family_df, unassigned_individuals)
}
\arguments{
\item{family_df}{Data frame. Data frame containing four columns: id, sex, momid, dadid}

\item{unassigned_individuals}{Data frame. Data frame containing id and sex}
}
\value{
List. List containing parents data frame, the remaining unassigned individuals data frame and assigned individuals data frame
}
\description{
pairs children with unassigned individuals (spouses)
}
