% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_specific_helper_functions.R
\name{parameter_score}
\alias{parameter_score}
\title{ML Score function}
\usage{
parameter_score(XtVinv, semi_def_matrix, omega_inv, Py, residual_vec)
}
\arguments{
\item{XtVinv}{the XtVinv matrix.}

\item{semi_def_matrix}{list of covariance component matrices.}

\item{omega_inv}{the inverse of omega as calculated by omega_func().}

\item{Py}{Py matrix as calculated by Py_func().}

\item{residual_vec}{vector of residuals as calculated by residual_function().}
}
\value{
Maximum likelihood score.
}
\description{
Function to calculate the maximum likelihood score.
}
