% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_kinship_data.R
\name{print.FDG}
\alias{print.FDG}
\title{Family data set generator}
\usage{
\method{print}{FDG}(x, ...)
}
\arguments{
\item{x}{an object of class FDG.}

\item{...}{additional arguments}
}
\value{
Print for object of class FDG.
}
\description{
Family data set generator
}
