% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fisher_scoring.R
\name{print.lmmgen}
\alias{print.lmmgen}
\title{Print lmmgen model likelihood estimates}
\usage{
\method{print}{lmmgen}(x, ...)
}
\arguments{
\item{x}{an object of class lmmgen}

\item{...}{additional arguments}
}
\value{
Print for object of class lmmgen
}
\description{
Print lmmgen model likelihood estimates
}
