% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_specific_helper_functions.R
\name{Py_func}
\alias{Py_func}
\title{Py matrix}
\usage{
Py_func(omega_inv, residual_vec)
}
\arguments{
\item{omega_inv}{the inverse of omega as calculated by omega_func().}

\item{residual_vec}{vector of residuals as calculated by residual_function().}
}
\value{
Py matrix
}
\description{
Py matrix
}
