% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_specific_helper_functions.R
\name{residual_function}
\alias{residual_function}
\title{Residual}
\usage{
residual_function(outcomes, design_matrix, beta)
}
\arguments{
\item{outcomes}{vector of outcomes.}

\item{design_matrix}{design matrix.}

\item{beta}{vector of beta coefficients.}
}
\value{
Residuals.
}
\description{
Calculates the residuals given observed values, beta coefficient and a design matrix.
}
