% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_func.R
\name{RSS_func}
\alias{RSS_func}
\title{Residual Sum of Squares}
\usage{
RSS_func(X, semi_def_matrix, y, params)
}
\arguments{
\item{X}{design matrix.}

\item{semi_def_matrix}{variance/covariance matrix.}

\item{y}{outcome vector.}

\item{params}{parameter vector.}
}
\value{
Residual sum of squares.
}
\description{
Calculates the residual sum of squares.
}
