% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fisher_scoring.R
\name{score_fisher_function}
\alias{score_fisher_function}
\title{Fisher scoring algorithm}
\usage{
score_fisher_function(
  design_matrix,
  semi_def_matrix,
  outcomes,
  fixed_effect_par,
  random_effect_par,
  small_value_threshold = 1e-12,
  add_small_constant = 1e-09
)
}
\arguments{
\item{design_matrix}{fixed effects design matrix for a single block.}

\item{semi_def_matrix}{list of covariance structure matrices for each block.}

\item{outcomes}{vector of outcomes for the block.}

\item{fixed_effect_par}{vector of initial fixed effect parameters.}

\item{random_effect_par}{vector of initial random effect parameters.}

\item{small_value_threshold}{values in matrices below this threshold is set to 0.}

\item{add_small_constant}{adds this constant to diagonal of matrices with small values for stability.}
}
\value{
List containing scores and fisher information.
}
\description{
Computes the relevant parts of the fisher scoring algorithm for a single block.
}
