% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_func.R
\name{XtVinvX_func}
\alias{XtVinvX_func}
\title{XtVinvX}
\usage{
XtVinvX_func(X, semi_def_matrix, y, params)
}
\arguments{
\item{X}{design matrix.}

\item{semi_def_matrix}{variance/covariance matrix.}

\item{y}{outcome vector.}

\item{params}{parameter vector.}
}
\value{
XtVinvX
}
\description{
XtVinvX
}
